requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
click>=8.0.0
pyyaml>=6.0
rich>=13.0.0
tqdm>=4.65.0
numpy>=1.24.0
pandas>=2.0.0
flask>=3.0.0
gunicorn>=21.2.0
